TEMP: suite: Add client class
This class is a wrapper for rfb::CConnection and replaces a real
client-server connection with a fake one that doesn't use a network
connection.

The Client can consume a DummyOutStream produced by a suite::Server by
simply copying the OutStream and setting it as its InStream.

This makes it possible to record or orchestrate a VNC session and
process it in non-real-time.