TEMP: suite: Add Benchmark class
The Benchmark class is responsible for configuring and initializing all
Encoders to be benchmarked. One Benchmark run can contain several
Servers with different settings.

It parses a session file and decodes all framebuffer updates and creates
an Image for each one.

Each image is sent to the Server which call Manager::writeUpdate(),
where the performance measurement takes place.