TEMP: Add BruteForce class
The BruteForce class makes it possible to run a benchmark which ensures
that the encoder with the best compression for a given rectangle will
always be used.

This means a compression ration upper bound can be calculated for the
EncodingManager. The class BruteForceEncoder can be extended/modified to
use a different predicate to determine what the "best" encoder is for a
given rectangle.